feat: notify slack from github webhook listeners
Updates all GitHub events to add functionality for returning a message
payload consistent with a Slack attachment.

Updates all existing GitHub listeners to send a slack chat post for the
event provided.

The Status listener also checks to see if a PR context is active and, if
so, queries the search API for any PRs matching the sha that produced
the status update. If so, it uses that data for the message, otherwise,
it assumes it is a generic build against a branch and sends a
notification of that.

Adds a GitHubReleaseSlackListener for notifying Slack of a new release.

Modifies the AttachmentColor class to only provide validation logic.

Modifies the Attachment class to allow arbitrary fields, and to validate
that specific fields are always present. It also now tests to see which
of the pretext and text fields are present, and adds a mrkdwn_in field
indicating all found to the attachment.

Makes the DiscoursePost event consistent with the various GitHub events,
and updates the DiscoursePostListener to have it create an Attachment
instance before attaching it to the ChatPostMessage.