feat: /builddocs command
Provides updates to the SlashCommandResponseFactory to allow delivering
an unauthorized response, as well as a more featureful response
containing blocks to users.

Adds a DocsBuildAction event under the GitHub namespace for describing a
docs-build dispatch event to send to a repository, and a listener that
accepts it and sends the event to GitHub.

Adds the BuildDocs slash command, which validates that only an
authorized user has triggered the command, and that the repository is
in an organization we manage. It then creates a DocsBuildAction and
dispatches it, before returning a response indicating it has queued the
request.