feat: /registerrepo command
Extracts GitHubClient, for making API requests to GitHub; reduces the
number of dependencies for most scenarios by two.

Extracts UrlHelper, which combines ServerUrlHelper + UrlHelper.

Adds the method SlackClientInterface::sendWebhookMessage for sending
messages via the response_url provided from a slash command, which goes
to a Slack webhook.

Extracts ValidateRepoArgumentTrait, for validating commands where
`{repo}` is the only argument.

Adds RegisterRepoCommand, which triggers a RegisterWebhook event, which
is in turn handled by the RegisterWebhookListener.