feat: adds /regenerate-tsc-list slash command
Refactors `AuthorizedUserList` to extract a `build()` command, which is
called during initialization, and can be called again later if needed.

Adds the /regenerate-tsc-list slash command via
`RegenerateAuthorizedUserListCommand`, which dispatches a
`RegenerateAuthorizedUserList` command, which is in turn listened for
via a `RegenerateAuthorizedUserListListener`. The listener calls on
`AuthorizedUserList::build()`, and reports the results.