refactor: extract HttpClientInterface
In most cases where we need an HTTP client, we need a `send()` method
that accepts a PSR-7 request and returns a PSR-7 response, and a PSR-17
request factory.

HttpClientInterface extends RequestFactoryInterface and adds the
`send()` signature.

HttpClient accepts a Guzzle client and a PSR-17 request factory instance
and proxies to them.

GitHubClient now decorates an HttpClientInterface instance (and also
implements the interface). SlackClient now proxies to an
HttpClientInterface instance.

Classes that were previously depending on both a guzzle client and a
request factory now instead depend on an HttpClientInterface.

This change will allow us to swap out HTTP client implementations in the
future if we so choose.