feat: re-enable Slack logging
To do this, I've switched to using the SlackWebhookHandler instead of
SlackHandler with Monolog. Per their own documentation, it's easier to
setup, and less prone to error (as SlackHandler uses the OAuth2
workflow, and thus tokens can expire).

For this to work, you MUST setup an incoming webhook on Slack for the
bot, and supply it in the SLACK_WEBHOOK_LOGGING env variable.