fix: ensure we correctly determine when we have an "OK" response from Slack
Slack always uses a 200 status code.

When indicating success, one of two things can happen:

- it can return the string `ok` as the full content of the response
- it can return `{"ok":true}` as the response content

As such, we need to detect both.

This also affects the "payload". Since we cannot assume the content can
be deserialized JSON, we have to set it to an empty array if it is not
one.

This change also modifies how responses from Slack are logged such that
they now include the response body, so we can determine what happened
more readily.