feat: separate logging for HTTP web server
Currently, all application and HTTP logging uses the same logger. This
means that web server logs for status codes >= 400 are going to Slack
currently, when they should not.

This patch modifies the application to provide a
`SwooleLoggerFactory::SWOOLE_LOGGER` service, pointing to a new
`App\Factory\AccessLogFactory`. This factory creates a new Monolog
logger using the name "http", and attaches the handler defined in
`mezzio-swoole.log_handler`.

In order to share handler definitions, I've added a dependency on
laminas-config-aggregator-parameters, and defined a default log handler
in the `config/config.php` file, in the ParameterPostProcessor
configuration. This is then re-used by both the
`mezzio-swoole.log_handler` and `monolog.handlers.default`
configuration.