fix: fields layout in slack messages
While the Slack documentation shows that fields are stacked 2 per row,
and demonstrates using this to align labels and values, in practice,
this fails as soon as you view it in a mobile app, as they are only
stacked 1 per row.

As it turns out, you CAN provide newlines, which means we can combine
the label and its value into a single field.