fix: ensure Discourse gets posted to Slack
I was pulling the wrong property from the Discourse webhook payload for
the body. The "raw" property does not exist, but the "cooked" one does.
Unfortunately, it contains HTML, which Slack does not allow. As such,
this patch also introduces a formatter for translating HTML to Slack
markup, and filters the "cooked" property through that when returning
the block content.