fix: always check if a status is related to a PR
We cannot assume we'll have PR information in the context. As such, we
need to look at commits to see if they are related to a pull request. As
such, I've removed the `isForPullRequest()` method from the
`GitHubStatus` class, and now always call `fetchPullRequestData()`
method to try and find a related PR.

Additionally, I discovered that the search query was malformed. It
requires a "q" parameter, with the search criteria, and should include
the criteria "type:pr". Finally, the values need to be formatted via
http_build_query to ensure they are escaped properly.