feat: full-fledged HTML to Slack formatter
The new approach uses DOMDocument to parse the HTML, using specific
element types to determine how to output the text as follows:

Element Type | Formatting
-------------------------
b, strong, h1 - h6 | `*%s*`
i, em | `_%s_`
li | `- %s`
pre | Surround with fences (```)
code | Surround with `

When an aside is encountered, it checks to see if it has the class
"onebox"; if so, it pulls the summary from its header.source link and
prints that, ignoring the rest. Otherwise, it treats it as any other
HTML tag (stripping it, and parsing the contents). This addresses
formatting coming from Discourse.

The functionality for rewriting the links to prepend the Discourse base
URL was removed, and will be moved to the DiscoursePost class itself.