feat: send tweets to #news channel on Slack
This patch completes the work for accepting a tweet from Twitter and
pushing it to Slack.

Incoming tweets are verified via:

- A shared secret provided in the URL.
- Validation that the body contents includes text, url, and timestamp
  elements.

The handler then creates an `App\Twitter\Tweet` instance from that data
and dispatches it via the event dispatcher.

The listener on that instance uses the data to push a message into the
configured slack channel (#news by default).

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>