fix: cast media ID to string after successful Twitter media upload
The return typehint for TweetListener::uploadMedia specifies a string or
null value, but twitter media and tweet IDs can be integers. Since they
will most likely run out of valid integers at some point, we have to
treat them as strings. This patch casts the value to a string on return.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>