fix: original username must be sent with replies
The Twitter API indicates that the username needs to be part of the
message sent when doing a reply. However, I found here:

- https://sung.codes/blog/2019/07/28/adding-a-twitter-reply-via-twitter-api/

that it can be sent as a parameter to statuses/update. I have added
logic to autodetect the originating user via the URL and pass it as a
parameter.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>