refactor: do not reply immediately to slash commands
Per the [Slack slash command documentation](https://api.slack.com/interactivity/slash-commands#responding_to_commands),
There are two ways to respond to a Slash command:

- Immediately, in the response returned to the request sent by Slack to
  the webhook.
- Later, in a request sent to the temporary webhook provided in the
  slash command response_url.

The two are mutually exclusive; if you send an immediate response, you
cannot later send a response via the webhook (it will be ignored).
However, if you opt to send via webhook, you MUST send an empty 200
response acknowledging receipt.

This patch modifies the `SlashCommandInterface::dispatch()` signature to
make the return type nullable (this is NOT a BC break, as
implementations are allowed to define a _narrower_ return value). This
allows us to choose per-command whether or not an immediate response
will be returned.

The method `createAcknowledgementResponse()` was added to the
`SlashCommandResponseFactory`; this method creates and returns an empty
200 response.

I have updated the `SlashCommands` class to call and return
`SlashCommandResponseFactory::createAcknowledgementResponse()` if a
`null` value is returned from a command.

I have updated all commands to return a null value at this time, as all
listeners they dispatch to send a response via webhook currently.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>