feat: update to PHP 8.1
In order to use Monolog 3.0, we need PHP 8.1.
This patch updates to PHP 8.1, which means updating to each of the following as well:

- Swoole 4.8.11
- mezzio/mezzio-swoole 4.3 (which allows us to also remove phly/phly-swoole-taskworker, as that functionality is baked in)
- mezzio/mezzio-tooling 2.5

I've also updated the Platform.sh config to use a PHP 8.1 image and Swoole 4.8.11, and now provide some additional `configure` options when building.

Tests pass, but I'll need to validate the webhooks.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>